# Add remove() to bitmask_binop!