In this commit, I am adding a new function remove() to the bitmask_binop! macro. This function removes some set bits from the self instance.